+++ pyannote/database/custom.py
@@ -27,6 +27,8 @@
 # Hervé BREDIN - http://herve.niderb.fr
 # Pavel KORSHUNOV - https://www.idiap.ch/~pkorshunov/
 
+
+import os
 from . import protocol as Protocol
 from .database import Database
 from .util import load_lst, load_uem, load_mdtm, load_rttm
@@ -34,10 +36,65 @@ import functools
 from pathlib import Path
 import yaml
 import pandas as pd
+from pyannote.core import Annotation, Timeline
 
 
 from . import DATABASES, TASKS
 
+
+# annoying mapping...
+SUBSET_MAPPING = {'train': 'trn', 'development': 'dev', 'test': 'tst'}
+
+
+def meta_subset_iter(config):
+    """This function will become a xxx_iter method of a meta-protocol
+
+
+    `config` comes from this part of ~/.pyannote/database.yml
+    (marked with arrows below)
+
+    ~~~~ ~/.pyannote/database.yml ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    Protocols:
+      X:
+        SpeakerDiarization:
+          ExtendedETAPE:
+            train:
+    -->       Etape.SpeakerDiarization.TV: [train]
+    -->       REPERE.SpeakerDiarization.Phase1: [train, development]
+    -->       REPERE.SpeakerDiarization.Phase2: [train, development]
+            development:
+              Etape.SpeakerDiarization.TV: [development]
+            test:
+              Etape.SpeakerDiarization.TV: [train]
+    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+    Parameters
+    ----------
+    config: `dict`
+        {protocol_name: list of subsets} dictionary.
+
+    Yields
+    ------
+    current_file : `dict`
+        Dictionary that provide information about a file.
+        This must include (at the very least) 'uri' and 'database' keys.
+    """
+
+    from . import get_protocol
+
+
+    # loop on all protocols
+    for protocol_name, subsets in config.items():
+
+        protocol = get_protocol(protocol_name)
+
+        # loop on requested subsets of current protocol
+        for subset in subsets:
+
+            # loop on all files of current subset
+            xxx_iter = getattr(protocol, f'{SUBSET_MAPPING[subset]}_iter')
+            for current_file in xxx_iter():
+                yield current_file
 
 
 def subset_iter(database_name, file_lst=None, file_rttm=None, file_uem=None):
@@ -82,32 +139,20 @@ def subset_iter(database_name, file_lst=None, file_rttm=None, file_uem=None):
     if file_lst is not None:
         uris = load_lst(file_lst)
 
-    # check URIs list consistency
-
-    # if file_lst is provided, make sure other files don't contain extra URIs
+    # when file_lst is provided, use this list of uris
     if len(uris) > 0:
+        pass
 
-        if not set(annotations).issubset(set(uris)):
-            msg = f'{file_rttm} contains URIs that are not in {file_lst}.'
-            raise ValueError(msg)
-
-        if not set(annotated).issubset(set(uris)):
-            msg = f'{file_rttm} contains URIs that are not in {file_lst}.'
-            raise ValueError(msg)
-
-    # if file_uem is provided, make sure file_rttm doesn't contain extra URIs
+    # when file_uem is provided, use this list of uris
     elif len(annotated) > 0:
         uris = sorted(annotated)
-        if not set(annotations).issubset(set(uris)):
-            msg = f'{file_rttm} contains URIs that are not in {file_uem}.'
-            raise ValueError(msg)
 
     # if file_rttm is the only file provided, use its list of URIs
     elif len(annotations) > 0:
         uris = sorted(annotations)
 
     # complain if we weren't able to get any list of URIs
-    if len(uris) == 0:
+    if not uris:
         msg = f'Empty protocol'
         raise ValueError(msg)
 
@@ -115,27 +160,43 @@ def subset_iter(database_name, file_lst=None, file_rttm=None, file_uem=None):
     for uri in uris:
 
         # initialize current file with the mandatory keys
-        current_file = {'database': database_name,
-                        'uri': uri}
+        current_file = {'database': database_name, 'uri': uri}
 
-        # add 'annotation' if/when available
-        if uri in annotations:
-            current_file['annotation'] = annotations[uri]
+        # add 'annotation' when RTTM file is provided
+        # defaults to empty Annotation because of
+        # github.com/pyannote/pyannote-database/pull/13#discussion_r261564520)
+        if file_rttm is not None:
+            current_file['annotation'] = annotations.get(
+                uri, Annotation(uri=uri))
 
-        # add 'annotated' if/when available
-        if uri in annotated:
-            current_file['annotated'] = annotated[uri]
+        # add 'annotated' when UEM file is provided
+        # defaults to empty Timeline for the same reason as above
+        if file_uem is not None:
+            current_file['annotated'] = annotated.get(uri, Timeline(uri=uri))
 
         yield current_file
 
 
-def add_custom_protocols(custom_yml=None):
-    """Update pyannote.database.{DATABASES, TASKS} with custom protocols
+def get_init(register):
+
+    def init(self, preprocessors={}):
+        super(self.__class__, self).__init__(preprocessors=preprocessors)
+        for protocol in register:
+            self.register_protocol(*protocol)
+
+    return init
+
+
+def add_custom_protocols(config_yml=None):
+    """Update pyannote.database.{DATABASES|TASKS} with custom & meta protocols
 
     Parameters
     ----------
-    custom_yml : `Path`, optional
-        path to YAML file with description of the database and its protocols
+    config_yml : `str`, optional
+        Path to pyannote.database configuration file in YAML format.
+        Defaults to the content of PYANNOTE_DATABASE_CONFIG environment
+        variable if defined and to "~/.pyannote/database.yml" otherwise.
+        Path to YAML file with description of the database and its protocols.
 
     Returns
     -------
@@ -143,24 +204,34 @@ def add_custom_protocols(custom_yml=None):
     pyannote.database.TASKS
     """
 
-    if custom_yml is None:
-        custom_yml = Path('~/.pyannote/custom.yml').expanduser()
+    if config_yml is None:
+        config_yml = os.environ.get("PYANNOTE_DATABASE_CONFIG",
+                                    "~/.pyannote/database.yml")
+    config_yml = Path(config_yml).expanduser()
 
     try:
-        with open(custom_yml, 'r') as fp:
-            custom_config = yaml.load(fp)
+        with open(config_yml, 'r') as fp:
+            config = yaml.load(fp)
 
     except FileNotFoundError:
-        custom_config = dict()
+        config = dict()
+
+    databases = config.get('Protocols', dict())
+
+    # make sure meta-protocols are processed last (relies on the fact that
+    # dicts are iterated in insertion order since Python 3.6)
+    x = databases.pop('X', None)
+    if x is not None:
+        databases['X'] = x
 
     # for each database
-    for database_name, dbtasks in custom_config.items():
+    for database_name, tasks in databases.items():
 
         # this list is meant to contain one class per protocol
         register = []
 
         # for each type of task
-        for task_name, protocols in dbtasks.items():
+        for task_name, protocols in tasks.items():
 
             # update TASKS dictionary
             if task_name not in TASKS:
@@ -169,14 +240,13 @@ def add_custom_protocols(custom_yml=None):
 
             # only speaker diarization TASKS are supported for now...
             if task_name != 'SpeakerDiarization':
-                msg = ('Only speaker diarization protocols are supported for now.')
+                msg = (
+                    'Only speaker diarization protocols are supported for now.'
+                )
                 raise ValueError(msg)
 
             # get protocol base class (here: SpeakerDiarizationProtocol)
             protocol_base_class = getattr(Protocol, f'{task_name}Protocol')
-
-            # annoying mapping...
-            SUBSET = {'train': 'trn', 'development': 'dev', 'test': 'tst'}
 
             # for each protocol
             for protocol_name, subsets in protocols.items():
@@ -185,41 +255,47 @@ def add_custom_protocols(custom_yml=None):
                 protocol_methods = {}
 
                 # for each subset
-                for subset, sub in SUBSET.items():
+                for subset, sub in SUBSET_MAPPING.items():
 
-                    # if protocols.yml does not provide a file for this subset,
+                    # if database.yml does not provide a file for this subset,
                     # skip it
                     if subset not in subsets:
                         continue
 
-                    paths = subsets[subset]
-                    file_rttm, file_lst, file_uem = None, None, None
-                    if 'annotation' in paths:
-                        file_rttm = Path(paths['annotation'])
-                    if 'uris' in paths:
-                        file_lst = Path(paths['uris'])
-                    if 'annotated' in paths:
-                        file_uem = Path(paths['annotated'])
+                    # special treatment for meta-protocols
+                    if database_name == 'X':
 
-                    # define xxx_iter method
-                    protocol_methods[f'{sub}_iter'] = functools.partial(
-                        subset_iter, database_name, file_rttm=file_rttm,
-                        file_lst=file_lst, file_uem=file_uem)
+                        protocol_methods[f'{sub}_iter'] = functools.partial(
+                            meta_subset_iter, subsets[subset])
 
-            # create protocol class on-the-fly
-            protocol = type(protocol_name, (protocol_base_class,), protocol_methods)
+                    else:
 
-            # keep track of this protocol -- database.__init__ needs to register it later...
-            register.append((task_name, protocol_name, protocol))
+                        paths = subsets[subset]
+                        file_rttm, file_lst, file_uem = None, None, None
+                        if 'annotation' in paths:
+                            file_rttm = Path(paths['annotation'])
+                        if 'uris' in paths:
+                            file_lst = Path(paths['uris'])
+                        if 'annotated' in paths:
+                            file_uem = Path(paths['annotated'])
+
+                        # define xxx_iter method
+                        protocol_methods[f'{sub}_iter'] = functools.partial(
+                            subset_iter, database_name, file_rttm=file_rttm,
+                            file_lst=file_lst, file_uem=file_uem)
+
+                # create protocol class on-the-fly
+                protocol = type(protocol_name, (protocol_base_class,),
+                                protocol_methods)
+
+                # keep track of this protocol -- database.__init__ needs to
+                # register it later...
+                register.append((task_name, protocol_name, protocol))
 
         # define Database.__init__ method
-        def db_init(self, preprocessors={}):
-            super(self.__class__, self).__init__(preprocessors=preprocessors)
-            for protocol in register:
-                self.register_protocol(*protocol)
 
         # create database class on-the-fly
         DATABASES[database_name] = type(database_name, (Database,),
-                                        {'__init__': db_init})
+                                        {'__init__': get_init(register)})
 
     return DATABASES, TASKS